New wsdl august4 (#495)
* 1. New jar built :
	a. New FareConvertCurrency API
	b. New Air Multiavailibility API
	c. New Upsell with and without PNR API's
	d. New FareFamilyDescription API

2. Implemented FareConvertCurrency API
3. Added Try Catch to fare rules
4. Improved Logging at Service Handler

* Removed indigo library due to conflicts from GDS as it is present in indigo service

* Reverted back to old db creds for development

+++ app/utils/RefundHelper.java
@@ -71,7 +71,7 @@ public class RefundHelper {
 
                             BigDecimal amount = taxType.getAmount();
                             String currency = taxType.getCurrencyCode();
-                            BigInteger decimalPlaces = BigInteger.valueOf(taxType.getDecimalPlaces());
+                            BigInteger decimalPlaces = taxType.getDecimalPlaces();
                             String category = taxType.getCategory();
                             String isoCode = taxType.getISOCode();
 
@@ -99,7 +99,7 @@ public class RefundHelper {
                             BigDecimal percent = penalty.getPercent();
                             BigDecimal amount = penalty.getAmount();
                             String currency = penalty.getCurrencyCode();
-                            BigInteger decimalPlaces = BigInteger.valueOf(penalty.getDecimalPlaces());
+                            BigInteger decimalPlaces = penalty.getDecimalPlaces();
                             String penaltyType = penalty.getPenaltyType();
 
                             detailedPenaltyInformation.setPercent(percent);
@@ -126,7 +126,7 @@ public class RefundHelper {
                             CommissionType.CommissionPayableAmount commissionPayableAmount = commission.getCommissionPayableAmount();
                             BigDecimal amount = commissionPayableAmount.getAmount();
                             String currency = commissionPayableAmount.getCurrencyCode();
-                            BigInteger decimalPlaces = BigInteger.valueOf(commissionPayableAmount.getDecimalPlaces());
+                            BigInteger decimalPlaces = commissionPayableAmount.getDecimalPlaces();
 
                             detailedCommissionInformation.setPercent(percent);
                             detailedCommissionInformation.setAmount(amount);
@@ -150,7 +150,7 @@ public class RefundHelper {
                             String qualifier = monetaryInformation.getQualifier();
                             BigDecimal amount = monetaryInformation.getAmount();
                             String currency = monetaryInformation.getCurrencyCode();
-                            BigInteger decimalPlaces = BigInteger.valueOf(monetaryInformation.getDecimalPlaces());
+                            BigInteger decimalPlaces = monetaryInformation.getDecimalPlaces();
 
                             detailedMonetaryInformation.setQualifier(qualifier);
                             detailedMonetaryInformation.setAmount(amount);
